+++ routes/promoRouter.js
@@ -21,6 +21,10 @@ promoRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser ,(req,res,next) =>{
+    if(!req.user.admin) {
+        var err = new Error('You are not authorized for this operation');
+        throw err;
+    }
     Promotions.create(req.body)
     .then((promo)=> {
         res.statusCode = 200;
@@ -36,6 +40,10 @@ promoRouter.route('/')
     res.end('PUT not supported on /promotions');
 })
 .delete(authenticate.verifyUser ,(req,res,next) =>{
+    if(!req.user.admin) {
+        var err = new Error('You are not authorized for this operation');
+        throw err;
+    }
     Promotions.deleteMany()
     .then((response) => {
         res.statusCode= 200;
@@ -62,11 +70,19 @@ promoRouter.route('/:promoId')
 })
 
 .post(authenticate.verifyUser ,(req,res,next) =>{
+    if(!req.user.admin) {
+        var err = new Error('You are not authorized for this operation');
+        throw err;
+    }
     res.statusCode = 403;
     res.end('POST not supported for /promos/'+ req.params.promoId);    
 })
 
 .put(authenticate.verifyUser ,(req,res,next) =>{
+    if(!req.user.admin) {
+        var err = new Error('You are not authorized for this operation');
+        throw err;
+    }
     Promotions.findByIdAndUpdate(req.params.promoId,{
         $set: req.body
     },{new: true})
@@ -81,6 +97,10 @@ promoRouter.route('/:promoId')
 })
 
 .delete(authenticate.verifyUser ,(req,res,next) =>{
+    if(!req.user.admin) {
+        var err = new Error('You are not authorized for this operation');
+        throw err;
+    }
     Promotions.findByIdAndRemove(req.params.promoId)
     .then((promos) => {
         res.statusCode = 200;
